Very messy update: Working on a problem with the membranes, specifically horizontalSqueeze. Have made it into its own kernel now, and provided rendering tools, to no avail yet. Next step is a fully async Display, so that will be next update.

+++ code/LIMA_BASE/include/LimaTypes.cuh
@@ -130,7 +130,7 @@ struct Float3 {
 	}
 	__host__ __device__ Float3 round() const { return Float3{ roundf(x), roundf(y), roundf(z) }; }
 	__host__ __device__ Float3 square() const { return Float3(x * x, y * y, z * z); }
-	__host__ __device__ inline float len() const { return (float)sqrtf(x * x + y * y + z * z); }
+	__host__ __device__ inline float len() const { return sqrtf(x * x + y * y + z * z); }
 	__host__ __device__ inline double len_d() const { return sqrt((double)x * x + (double)y * y + (double)z * z); }
 	__host__ __device__ inline float lenSquared() const { return (float)(x * x + y * y + z * z); }
 	__host__ __device__ Float3 zeroIfAbove(float a) { return Float3(x * (x < a), y * (y < a), z * (z < a)); }
